Incorrect code verification
I implemented an if statement to verify whether the entered code matches the predefined correct code (correctCode). If the codes match, a success modal is displayed, and the user is navigated to the login screen. If the codes do not match, an error modal is displayed to notify the user.